// src/models/spt/ITemplateItem.ts
export interface ITemplateItem {
  _id: string;
  _name: string;
  _parent: string;
  _type: ItemType;
  _props: IProps;
  _proto?: string;
}

export enum ItemType {
  NODE = "Node",
  ITEM = "Item",
}

export interface IProps {
// ...
  type?: string;
// ...
}